Change trigger wheel ISR
In order to further shorten the trigger wheel ISR, a new volatile
variable Tooth_Detected was created. This variable is set to true when a
tooth is detected. The main loop function will acknowledge tooth
detection, then immediately set the variable back to false. Tooth number
calculation and gap detection was moved to the main loop function,
further reducing the need for volatile variables and calculations in the
ISR.