Correct VE table interpolation issue
Before, there was an issue with VE table interpolation wherein it was
possible for lookup lower and upper indices for MAP or engine speed to
be equal, leading to division by zero issues. This would occur when MAP
or RPM exceeded maximum tabulated value. By changing the upper limit for
lower index to number of tabulated values minus two from number of
tabulated values minus one, the problem is circumvented.